Rk356x: UEFI eMMC support
The eMMC controller is mostly compatible to standard SDHC.  There are
three quirks that we can override easily using SdMmcOverride protocol.

One issue is that toggling POWER_CTRL off results in SDCLK_ENABLE being
turned off as well.  We need to re-enable the bit to be able to talk to
the card.

Another issue is that the divisor is broken, and instead the CCLK_EMMC
clock should be configured to the target frequency.  While there we can
also setup the DLL correctly.

The last issue is that HS400 needs a different setting compared to other
controllers.