ROC-RK3566-PC: add support for WiFi chips on second SD controller
This provides the ACPI table information for the second MSHC on the
RK3566.  This one is commonly used to connect WiFi SDIO cards.  Since
not all boards have it routed, make the controller optional and only
enable it on the ROK-RK3566-PC.

Initialize the WiFi and MSHC controller by configuring the IOMUX, the
SDMMC1 clock and power cycling the WiFi card.

Reorder the PCD identifier to leave some gap between the components.