SOQuartz: sync with Quartz64 & ROC-RK3566-PC
Since SOQuartz was copied from Quartz64, a few things have changed in
there.  Like the Quartz64, it also has a YT8511C PHY with the same
GPIO pins for reset and TX/RX delays.  It might be a sensible idea to
deduplicate the Phy init code.  The PMIC on the other hand is the RK809
as implemented on the ROC-RK3566-PC, so copy that code to make sure that
the PMIC is turned on fully to make HDMI work.