SOQuartz: update GPIO pins
The GPIO pins copied from Quartz64 are not correct.  The WORK_LED is on
a different pin, and it needs to be driven low.  There is no GPIO pin to
enable USB host power, only the USB data pins are routed to the carrier.
There is no PCIe power pin, so remove those settings.  Unfortunately we
also have to change the PCIe host code, as passing the define being set
to 0xffffffff to a function that only takes UINT8 makes the compiler
error out.